tests/wpt/css-tests/css-exclusions-1_dev/xhtml1/wrap-flow-002.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
        <title>CSS Exclusions Test: wrap-flow property set to start</title>
        <link href="mailto:jacobg@adobe.com" rel="author" title="Jacob Goldstein" />
        <link href="http://www.w3.org/TR/css3-exclusions/#wrap-flow-property" rel="help" />
        <meta content="ahem dom" name="flags" />
        <meta content="Text should flow from the beginning of the line to the exclusion, as well as above and below" name="assert" />
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="resources/helper.js"></script>
        <style>
            * {
                margin: 0px;
                padding: 0px;
            }
            .exclusion {
                wrap-flow: start;
                position: absolute;
                top: 40px;
                left: 40px;
                width: 60px;
                height: 25px;
                background: blue;
            }
            #content {
                width: 400px;
                line-height: 12px;
                font: 12px Ahem;
            }
        </style>
    </head>
    <body>
        <div class="exclusion"></div>

        <div id="content">
            <span id="linesAbove">ethical<br />beer<br />craft<br /></span>
            <span id="lineLeft1">foo</span>
            <span id="lineLeft2">sed</span>
            <span id="linesBelow">Vegan<br />cliche<br />retro</span>
        </div>

        <div id="log"></div>

        <script type="text/javascript">
            test(function() {assert_equals(checkLinePos("lineLeft1",36,"top"), true)}, "Verify top of the 'lineLeft1' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("lineLeft2",48,"top"), true)}, "Verify top of the 'lineLeft2' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("lineLeft2",36,"right"), true)}, "Verify right of the 'lineLeft2' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("linesBelow",65,"top"), true)}, "Verify top of the 'linesBelow' span is positioned correctly");
        </script>
    

</body></html>
